moved VLT pup/apod/lyot routines to phase_amplitude_functions.py
